Formated code with prettier.

+++ src/app/api/users/profile/delete/route.ts
@@ -6,14 +6,12 @@ import { authOptions } from "../../../../../utils/authOptions";
 import { eq } from "drizzle-orm";
 
 export async function DELETE() {
+  const session = await getServerSession(authOptions);
 
-    const session = await getServerSession(authOptions);
-
-    try {
-        await db.delete(profiles)
-                .where(eq(profiles.user_id, session.user.id));
-        return NextResponse.json("Profile sucsessfully deleted");
-    } catch (error) {
-        return NextResponse.json(error);
-    }
+  try {
+    await db.delete(profiles).where(eq(profiles.user_id, session.user.id));
+    return NextResponse.json("Profile sucsessfully deleted");
+  } catch (error) {
+    return NextResponse.json(error);
+  }
 }
